Create circle.py
Here what I understand is the Shape class is the parent class and the  abstract method named draw() class is act as a blueprint for all other child classes but these child classes can have their own behavior. From the class method named create() looked up the requested shape in the _registry and returned a object of that class.